Add paginated food table, Meals form inputs format
